Polished implementation of causal clustering.
