+++ core/src/main/java/org/neo4j/ogm/session/SessionFactory.java
@@ -15,9 +15,9 @@ package org.neo4j.ogm.session;
 
 
 import org.neo4j.ogm.MetaData;
+import org.neo4j.ogm.autoindex.AutoIndexManager;
 import org.neo4j.ogm.config.Configuration;
 import org.neo4j.ogm.service.Components;
-import org.neo4j.ogm.autoindex.AutoIndexManager;
 
 /**
  * Used to create {@link Session} instances for interacting with Neo4j.
@@ -28,104 +28,101 @@ import org.neo4j.ogm.autoindex.AutoIndexManager;
  */
 public class SessionFactory {
 
-    private final MetaData metaData;
+	private final MetaData metaData;
 	private final AutoIndexManager autoIndexManager;
 
-    /**
-     * Constructs a new {@link SessionFactory} by initialising the object-graph mapping meta-data from the given list of domain
-     * object packages.
-     * <p>
-     * The package names passed to this constructor should not contain wildcards or trailing full stops, for example,
-     * "org.springframework.data.neo4j.example.domain" would be fine.  The default behaviour is for sub-packages to be scanned
-     * and you can also specify fully-qualified class names if you want to cherry pick particular classes.
-     * </p>
-     * Indexes will also be checked or built if configured.
-     *
-     * @param packages The packages to scan for domain objects
-     */
-    public SessionFactory(String... packages) {
-        this.metaData = new MetaData(packages);
+	/**
+	 * Constructs a new {@link SessionFactory} by initialising the object-graph mapping meta-data from the given list of domain
+	 * object packages.
+	 * <p>
+	 * The package names passed to this constructor should not contain wildcards or trailing full stops, for example,
+	 * "org.springframework.data.neo4j.example.domain" would be fine.  The default behaviour is for sub-packages to be scanned
+	 * and you can also specify fully-qualified class names if you want to cherry pick particular classes.
+	 * </p>
+	 * Indexes will also be checked or built if configured.
+	 *
+	 * @param packages The packages to scan for domain objects
+	 */
+	public SessionFactory(String... packages) {
+		this.metaData = new MetaData(packages);
 		this.autoIndexManager = new AutoIndexManager(this.metaData, Components.driver());
 		this.autoIndexManager.build();
-    }
+	}
 
-    /**
-     * Constructs a new {@link SessionFactory} by initialising the object-graph mapping meta-data from the given list of domain
-     * object classes.
-     * <p>
-     * This will only load the classes explicitly listed. No other classes will be loaded.
-     * </p>
-     * Indexes will also be checked or built if configured.
-     *
-     * @param classes The classes to load as domain objects
-     */
-    public SessionFactory(Class... classes) {
-        this.metaData = new MetaData(classes);
-        this.autoIndexManager = new AutoIndexManager(this.metaData, Components.driver());
-        this.autoIndexManager.build();
-    }
+	/**
+	 * Constructs a new {@link SessionFactory} by initialising the object-graph mapping meta-data from the given list of domain
+	 * object classes.
+	 * <p>
+	 * This will only load the classes explicitly listed. No other classes will be loaded.
+	 * </p>
+	 * Indexes will also be checked or built if configured.
+	 *
+	 * @param classes The classes to load as domain objects
+	 */
+	public SessionFactory(Class... classes) {
+		this.metaData = new MetaData(classes);
+		this.autoIndexManager = new AutoIndexManager(this.metaData, Components.driver());
+		this.autoIndexManager.build();
+	}
 
-    /**
-     * Constructs a new {@link SessionFactory} by initialising the object-graph mapping meta-data from the given list of domain
-     * object packages, and also sets the configuration to be used.
-     * <p>
-     * The package names passed to this constructor should not contain wildcards or trailing full stops, for example,
-     * "org.springframework.data.neo4j.example.domain" would be fine.  The default behaviour is for sub-packages to be scanned
-     * and you can also specify fully-qualified class names if you want to cherry pick particular classes.
-     * </p>
-     * Indexes will also be checked or built if configured.
-     *
-     * @param configuration The configuration to use
-     * @param packages The packages to scan for domain objects
-     */
-    public SessionFactory(Configuration configuration, String... packages) {
-        Components.configure(configuration);
-        this.metaData = new MetaData(packages);
-        this.autoIndexManager = new AutoIndexManager(this.metaData, Components.driver());
-        this.autoIndexManager.build();
-    }
+	/**
+	 * Constructs a new {@link SessionFactory} by initialising the object-graph mapping meta-data from the given list of domain
+	 * object packages, and also sets the configuration to be used.
+	 * <p>
+	 * The package names passed to this constructor should not contain wildcards or trailing full stops, for example,
+	 * "org.springframework.data.neo4j.example.domain" would be fine.  The default behaviour is for sub-packages to be scanned
+	 * and you can also specify fully-qualified class names if you want to cherry pick particular classes.
+	 * </p>
+	 * Indexes will also be checked or built if configured.
+	 *
+	 * @param configuration The configuration to use
+	 * @param packages The packages to scan for domain objects
+	 */
+	public SessionFactory(Configuration configuration, String... packages) {
+		Components.configure(configuration);
+		this.metaData = new MetaData(packages);
+		this.autoIndexManager = new AutoIndexManager(this.metaData, Components.driver());
+		this.autoIndexManager.build();
+	}
 
-    /**
-     * Constructs a new {@link SessionFactory} by initialising the object-graph mapping meta-data from the given list of domain
-     * object classes, and also sets the configuration to be used.
-     * <p>
-     * This will only load the classes explicitly listed. No other classes will be loaded.
-     * </p>
-     * Indexes will also be checked or built if configured.
-     *
-     * @param configuration The configuration to use
-     * @param classes The classes to load as domain objects
-     */
-    public SessionFactory(Configuration configuration, Class... classes) {
-        Components.configure(configuration);
-        this.metaData = new MetaData(classes);
-        this.autoIndexManager = new AutoIndexManager(this.metaData, Components.driver());
-        this.autoIndexManager.build();
-    }
+	/**
+	 * Constructs a new {@link SessionFactory} by initialising the object-graph mapping meta-data from the given list of domain
+	 * object classes, and also sets the configuration to be used.
+	 * <p>
+	 * This will only load the classes explicitly listed. No other classes will be loaded.
+	 * </p>
+	 * Indexes will also be checked or built if configured.
+	 *
+	 * @param configuration The configuration to use
+	 * @param classes The classes to load as domain objects
+	 */
+	public SessionFactory(Configuration configuration, Class... classes) {
+		Components.configure(configuration);
+		this.metaData = new MetaData(classes);
+		this.autoIndexManager = new AutoIndexManager(this.metaData, Components.driver());
+		this.autoIndexManager.build();
+	}
 
-    /**
-     * Retrieves the meta-data that was built up when this {@link SessionFactory} was constructed.
-     *
-     * @return The underlying {@link MetaData}
-     */
-    public MetaData metaData() {
-        return metaData;
-    }
+	/**
+	 * Retrieves the meta-data that was built up when this {@link SessionFactory} was constructed.
+	 *
+	 * @return The underlying {@link MetaData}
+	 */
+	public MetaData metaData() {
+		return metaData;
+	}
 
-    /**
-     * Opens a new Neo4j mapping {@link Session} using the Driver specified in the OGM configuration
-     * The driver should be configured to connect to the database using the appropriate
-     * DriverConfig
-     *
-     * @return A new {@link Session}
-     */
-    public Session openSession() {
-        if (Components.driver().getConfiguration().getDriverClassName().equals("org.neo4j.ogm.drivers.bolt.driver.BoltDriver")) {
-            return new Neo4jBoltSession(metaData, Components.driver());
-        }
-        return new Neo4jSession(metaData, Components.driver());
-    }
+	/**
+	 * Opens a new Neo4j mapping {@link Session} using the Driver specified in the OGM configuration
+	 * The driver should be configured to connect to the database using the appropriate
+	 * DriverConfig
+	 *
+	 * @return A new {@link Session}
+	 */
+	public Session openSession() {
+		return new Neo4jSession(metaData, Components.driver());
+	}
 
-    public void close() {
-    }
+	public void close() {
+	}
 }
