+++ core/src/main/java/org/neo4j/ogm/session/transaction/DefaultTransactionManager.java
@@ -16,7 +16,6 @@ package org.neo4j.ogm.session.transaction;
 import org.neo4j.ogm.driver.Driver;
 import org.neo4j.ogm.exception.TransactionManagerException;
 import org.neo4j.ogm.service.Components;
-import org.neo4j.ogm.session.Neo4jBoltSession;
 import org.neo4j.ogm.session.Neo4jSession;
 import org.neo4j.ogm.session.Session;
 import org.neo4j.ogm.transaction.AbstractTransaction;
@@ -29,157 +28,142 @@ import org.neo4j.ogm.transaction.TransactionManager;
  */
 public class DefaultTransactionManager implements TransactionManager {
 
-    private final Driver driver;
-    private final Session session;
+	private final Driver driver;
+	private final Session session;
 
-    private static final ThreadLocal<Transaction> TRANSACTION_THREAD_LOCAL = new ThreadLocal<>();
+	private static final ThreadLocal<Transaction> TRANSACTION_THREAD_LOCAL = new ThreadLocal<>();
 
-    public DefaultTransactionManager(Session session) {
-        this( session, Components.driver() );
-    }
+	public DefaultTransactionManager(Session session) {
+		this(session, Components.driver());
+	}
 
-    public DefaultTransactionManager(Session session, Driver driver) {
-        this.driver = driver;
-        this.driver.setTransactionManager(this);
-        this.session = session;
+	public DefaultTransactionManager(Session session, Driver driver) {
+		this.driver = driver;
+		this.driver.setTransactionManager(this);
+		this.session = session;
 
-        TRANSACTION_THREAD_LOCAL.remove();
-    }
+		TRANSACTION_THREAD_LOCAL.remove();
+	}
 
-    /**
-     * Opens a new TRANSACTION_THREAD_LOCAL against a database instance.
-     *
-     * Instantiation of the TRANSACTION_THREAD_LOCAL is left to the driver
-     *
-     * @return a new {@link Transaction}
-     */
-    public Transaction openTransaction() {
-        AbstractTransaction tx = ((AbstractTransaction) TRANSACTION_THREAD_LOCAL.get());
-        if (tx == null) {
-            return openTransaction(Transaction.Type.READ_WRITE);
-        } else {
-            return openTransaction(tx.type());
-        }
-    }
+	/**
+	 * Opens a new TRANSACTION_THREAD_LOCAL against a database instance.
+	 * Instantiation of the TRANSACTION_THREAD_LOCAL is left to the driver
+	 *
+	 * @return a new {@link Transaction}
+	 */
+	public Transaction openTransaction() {
+		AbstractTransaction tx = ((AbstractTransaction) TRANSACTION_THREAD_LOCAL.get());
+		if (tx == null) {
+			return openTransaction(Transaction.Type.READ_WRITE);
+		} else {
+			return openTransaction(tx.type());
+		}
+	}
 
-    /**
-     * Opens a new TRANSACTION_THREAD_LOCAL against a database instance.
-     *
-     * Instantiation of the TRANSACTION_THREAD_LOCAL is left to the driver
-     *
-     * @return a new {@link Transaction}
-     */
-    public Transaction openTransaction(Transaction.Type type) {
-        if (TRANSACTION_THREAD_LOCAL.get() == null) {
+	/**
+	 * Opens a new TRANSACTION_THREAD_LOCAL against a database instance.
+	 * Instantiation of the TRANSACTION_THREAD_LOCAL is left to the driver
+	 *
+	 * @return a new {@link Transaction}
+	 */
+	public Transaction openTransaction(Transaction.Type type) {
+		if (TRANSACTION_THREAD_LOCAL.get() == null) {
+			TRANSACTION_THREAD_LOCAL.set(driver.newTransaction(type, session.getLastBookmark()));
+		} else {
+			((AbstractTransaction) TRANSACTION_THREAD_LOCAL.get()).extend(type);
+		}
+		return TRANSACTION_THREAD_LOCAL.get();
+	}
 
-            String bookmark = null;
-            if (session instanceof Neo4jBoltSession) {
-                bookmark = ((Neo4jBoltSession)session).getLastBookmark();
-            }
+	/**
+	 * Rolls back the specified TRANSACTION_THREAD_LOCAL.
+	 * The actual job of rolling back the TRANSACTION_THREAD_LOCAL is left to the relevant driver. if
+	 * this is successful, the TRANSACTION_THREAD_LOCAL is detached from this thread. Any new objects
+	 * are reset in the session, so that their ids are reset to null.
+	 * If the specified TRANSACTION_THREAD_LOCAL is not the correct one for this thread, throws an exception
+	 *
+	 * @param transaction the TRANSACTION_THREAD_LOCAL to rollback
+	 */
+	public void rollback(Transaction transaction) {
+		if (transaction != getCurrentTransaction()) {
+			throw new TransactionManagerException("Transaction is not current for this thread");
+		}
 
-            TRANSACTION_THREAD_LOCAL.set(driver.newTransaction(type, bookmark));
-        } else {
-            ((AbstractTransaction) TRANSACTION_THREAD_LOCAL.get()).extend(type);
-        }
-        return TRANSACTION_THREAD_LOCAL.get();
-    }
+		for (Object object : ((AbstractTransaction) transaction).registeredNew()) {
+			((Neo4jSession) session).context().reset(object);
+		}
 
-    /**
-     * Rolls back the specified TRANSACTION_THREAD_LOCAL.
-     *
-     * The actual job of rolling back the TRANSACTION_THREAD_LOCAL is left to the relevant driver. if
-     * this is successful, the TRANSACTION_THREAD_LOCAL is detached from this thread. Any new objects
-     * are reset in the session, so that their ids are reset to null.
-     *
-     * If the specified TRANSACTION_THREAD_LOCAL is not the correct one for this thread, throws an exception
-     *
-     * @param transaction the TRANSACTION_THREAD_LOCAL to rollback
-     */
-    public void rollback(Transaction transaction) {
-        if (transaction != getCurrentTransaction()) {
-            throw new TransactionManagerException("Transaction is not current for this thread");
-        }
+		TRANSACTION_THREAD_LOCAL.remove();
+	}
 
-        for (Object object : ((AbstractTransaction) transaction).registeredNew()) {
-            ((Neo4jSession) session).context().reset( object );
-        }
+	/**
+	 * Commits the specified TRANSACTION_THREAD_LOCAL.
+	 * The actual job of committing the TRANSACTION_THREAD_LOCAL is left to the relevant driver. if
+	 * this is successful, the TRANSACTION_THREAD_LOCAL is detached from this thread.
+	 * If the specified TRANSACTION_THREAD_LOCAL is not the correct one for this thread, throws an exception
+	 *
+	 * @param tx the TRANSACTION_THREAD_LOCAL to commit
+	 */
+	public void commit(Transaction tx) {
+		if (tx != getCurrentTransaction()) {
+			throw new TransactionManagerException("Transaction is not current for this thread");
+		}
+		TRANSACTION_THREAD_LOCAL.remove();
+	}
 
-        TRANSACTION_THREAD_LOCAL.remove();
-    }
+	/**
+	 * Returns the current TRANSACTION_THREAD_LOCAL for this thread, or null if none exists
+	 *
+	 * @return this thread's TRANSACTION_THREAD_LOCAL
+	 */
+	public Transaction getCurrentTransaction() {
+		return TRANSACTION_THREAD_LOCAL.get();
+	}
 
-    /**
-     * Commits the specified TRANSACTION_THREAD_LOCAL.
-     *
-     * The actual job of committing the TRANSACTION_THREAD_LOCAL is left to the relevant driver. if
-     * this is successful, the TRANSACTION_THREAD_LOCAL is detached from this thread.
-     *
-     * If the specified TRANSACTION_THREAD_LOCAL is not the correct one for this thread, throws an exception
-     *
-     * @param tx the TRANSACTION_THREAD_LOCAL to commit
-     */
-    public void commit(Transaction tx) {
-        if (tx != getCurrentTransaction()) {
-            throw new TransactionManagerException("Transaction is not current for this thread");
-        }
-        TRANSACTION_THREAD_LOCAL.remove();
-    }
+	public boolean canCommit() {
 
-    /**
-     * Returns the current TRANSACTION_THREAD_LOCAL for this thread, or null if none exists
-     *
-     * @return this thread's TRANSACTION_THREAD_LOCAL
-     */
-    public Transaction getCurrentTransaction() {
-        return TRANSACTION_THREAD_LOCAL.get();
-    }
+		if (getCurrentTransaction() == null) {
+			return false;
+		}
 
-    public boolean canCommit()  {
+		AbstractTransaction tx = (AbstractTransaction) getCurrentTransaction();
 
-        if (getCurrentTransaction() == null) {
-            return false;
-        }
+		if (tx.extensions() == 0) {
+			if (tx.status() == Transaction.Status.COMMIT_PENDING || tx.status() == Transaction.Status.OPEN || tx.status() == Transaction.Status.PENDING) {
+				return true;
+			}
+		}
+		return false;
+	}
 
-        AbstractTransaction tx = (AbstractTransaction) getCurrentTransaction();
+	public boolean canRollback() {
 
-        if (tx.extensions() == 0) {
-            if (tx.status() == Transaction.Status.COMMIT_PENDING || tx.status() == Transaction.Status.OPEN || tx.status() == Transaction.Status.PENDING) {
-                return true;
-            }
-        }
-        return false;
-    }
+		if (getCurrentTransaction() == null) {
+			return false;
+		}
 
-    public boolean canRollback()  {
+		AbstractTransaction tx = (AbstractTransaction) getCurrentTransaction();
 
-        if (getCurrentTransaction() == null) {
-            return false;
-        }
+		if (tx.extensions() == 0) {
+			if (tx.status() == Transaction.Status.ROLLBACK_PENDING || tx.status() == Transaction.Status.COMMIT_PENDING || tx.status() == Transaction.Status.OPEN || tx.status() == Transaction.Status.PENDING) {
+				return true;
+			}
+		}
+		return false;
+	}
 
-        AbstractTransaction tx = (AbstractTransaction) getCurrentTransaction();
+	@Override
+	public void bookmark(String bookmark) {
+		session.lastBookmark(bookmark);
+	}
 
-        if (tx.extensions() == 0) {
-            if (tx.status() == Transaction.Status.ROLLBACK_PENDING || tx.status() == Transaction.Status.COMMIT_PENDING || tx.status() == Transaction.Status.OPEN || tx.status() == Transaction.Status.PENDING) {
-                return true;
-            }
-        }
-        return false;
-    }
+	// this is for testing purposes only
+	public void reinstate(AbstractTransaction tx) {
+		tx.reOpen();
+		TRANSACTION_THREAD_LOCAL.set(tx);
+	}
 
-    @Override
-    public void bookmark(String bookmark) {
-        if (session instanceof Neo4jBoltSession) {
-            ((Neo4jBoltSession)session).lastBookmark(bookmark);
-        }
-
-    }
-
-    // this is for testing purposes only
-    public void reinstate(AbstractTransaction tx) {
-        tx.reOpen();
-        TRANSACTION_THREAD_LOCAL.set(tx);
-    }
-
-    public void clear() {
-        TRANSACTION_THREAD_LOCAL.remove();
-    }
+	public void clear() {
+		TRANSACTION_THREAD_LOCAL.remove();
+	}
 }
